Adding lots of comments to the code, changing a very small part of the code
Major part of the code that I am going to deal with commented.

Also in the file hashed operation I changed the part related to "AddableOp" instace which is dealing with "Negate"

Nhan, Please check this to see if its correct or not. I fill this should be written like this.

Co-Authored-By: Nhan Thai <dainhan605@gmail.com>